变更数据包类型 uint16=>uint32 变更连接 id uint32=>uint64

+++ io/network/packet/base.go
@@ -5,18 +5,18 @@ import (
 )
 
 // 注册包处理动作编码
-var Packets = make(map[uint16]interface{})
+var Packets = make(map[uint32]interface{})
 
 // 反注册包处理动作编码
-var RPackets = make(map[interface{}]uint16)
+var RPackets = make(map[interface{}]uint32)
 
 // 包服务注册应由引用源程序处理并注册到当前包的packets之下
-func Register(op uint16, packet interface{}) {
+func Register(op uint32, packet interface{}) {
 	Packets[op] = packet
 	RPackets[reflect.TypeOf(packet)] = op
 }
 
-func OpCode(packet interface{}) uint16 {
+func OpCode(packet interface{}) uint32 {
 
 	elem := reflect.TypeOf(packet)
 
@@ -26,7 +26,7 @@ func OpCode(packet interface{}) uint16 {
 	return 0
 }
 
-func Packet(op uint16) interface{} {
+func Packet(op uint32) interface{} {
 	if v, ok := Packets[op]; ok {
 		return v
 	}
